requirements.txt
numpy==1.26.4
pandas==2.2.2
pydantic==2.7.1
scipy==1.11.4
scikit-learn==1.3.2
ccxt==4.3.69
pyyaml==6.0.1
click==8.1.7
loguru==0.7.2
redis==5.0.4
pyarrow==14.0.2
streamlit==1.35.0
plotly==5.20.0
python-dotenv==1.0.1
requests==2.31.0
vaderSentiment==3.3.2
tenacity==8.2.3
